extract game info to standalone component
